+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -58,7 +58,9 @@ export enum BroadcastEvent {
     UpdateConversationDataForTelemetry = "UpdateConversationDataForTelemetry",
     ContactIdNotFound = "ContactIdNotFound",
     SyncMinimize = "SyncMinimize",
-    OnWidgetError = "OnWidgetError"
+    OnWidgetError = "OnWidgetError",
+    FMLTrackingCompletedAck = "FMLTrackingCompletedAck",
+    FMLTrackingCompleted = "FMLTrackingCompleted"
 }
 
 // Events being logged
@@ -209,6 +211,8 @@ export enum TelemetryEvent {
     MessageSentCompleted = "MessageSentCompleted",
     MessageReceivedCompleted = "MessageReceivedCompleted",
     MessageLapTrack = "MessageLapTrack",
+    BotFirstMessageLapTrack = "BotFirstMessageLapTrack",
+    BotFirstMessageLapTrackError = "BotFirstMessageLapTrackError",
     MessageStartLapTrackError = "MessageStartLapTrackError",
     MessageStopLapTrackError = "MessageStopLapTrackError",
     SystemMessageReceivedCompleted = "SystemMessageReceivedCompleted",